Add a button, two checkmarks, and tab capability
Added the clear button and two checkmarks, the size of parseButton
and clearButton is now adjusted to fit the two checkboxes. Those
checkboxes are for the autocopy and for the overwrite option.

Also when the user press the tab key on the input, it will automatically
focus their selection to the parseButton. This behaviour is hard-coded
and cannot be changed.